+++ cmd/insprd/memory/tree/manager.go
@@ -16,6 +16,7 @@ var logger *zap.Logger
 // have been initialized
 func init() {
 	logger, _ = zap.NewProduction(zap.Fields(zap.String("section", "memory-tree")))
+	// logger, _ = zap.NewDevelopment(zap.Fields(zap.String("section", "memory-tree")))
 	// logger = zap.NewNop()
 }