+++ cmd/insprd/operators/operator.go
@@ -1,11 +1,23 @@
 package operators
 
 import (
+	"go.uber.org/zap"
 	"inspr.dev/inspr/cmd/insprd/memory"
 	"inspr.dev/inspr/cmd/insprd/memory/tree"
 	"inspr.dev/inspr/cmd/insprd/operators/nodes"
 	"inspr.dev/inspr/pkg/auth"
 )
+
+var logger *zap.Logger
+
+// init is called after all the variable declarations in the package have evaluated
+// their initializers, and those are evaluated only after all the imported packages
+// have been initialized
+func init() {
+	logger, _ = zap.NewProduction(zap.Fields(zap.String("section", "memory-tree")))
+	// logger, _ = zap.NewDevelopment(zap.Fields(zap.String("section", "operators")))
+	// logger = zap.NewNop()
+}
 
 // Operator is an operator for creating channels and nodes inside kubernetes
 // that communicate via Sidecars. The operators need two environment variables
@@ -17,11 +29,13 @@ type Operator struct {
 
 // Nodes returns the nodes that communicate via sidecars inside kubernetes
 func (op *Operator) Nodes() NodeOperatorInterface {
+	logger.Info("summoning Nodes Operator")
 	return op.nodes
 }
 
 // Channels returns the Channels Operator Interface for a given node
 func (op *Operator) Channels() ChannelOperatorInterface {
+	logger.Info("summoning Channels Operator")
 	return op.channels
 }